# storage/app/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'DOMPET Backend API API Documentation'
  description: 'Dokumentasi endpoint API untuk backend DOMPET/Zaku.'
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Dapatkan token dari endpoint <code>POST /api/auth/login</code>, lalu kirim sebagai Bearer token.'
security:
  -
    default: []
paths:
...
  /api/auth/verify-email:
    post:
      summary: 'Verify email'
      operationId: verifyEmail
      description: 'Memvalidasi kode verifikasi dan menandai email user sebagai terverifikasi.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Email verified successfully'
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Email verified successfully'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Verification code has expired.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Verification code has expired.'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  description: 'ID user yang akan diverifikasi.'
                  example: 1
                code:
                  type: string
                  description: 'Kode verifikasi 6 digit dari email/log.'
                  example: '123456'
              required:
                - user_id
                - code
      security: []
...
  /api/auth/forgot-password:
    post:
      summary: 'Forgot password'
      operationId: forgotPassword
      description: 'Membuat token reset password dan mengirimkannya ke email user.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Password reset link sent to your email'
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Password reset link sent to your email'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Email user.'
                  example: user@example.com
              required:
                - email
      security: []
  /api/auth/change-password:
    post:
      summary: 'Change password'
      operationId: changePassword
      description: 'Mengubah password user yang sedang login. Endpoint ini membutuhkan JWT Bearer token.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Password changed successfully'
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Password changed successfully'
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Authorization token is required.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Authorization token is required.'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  description: 'Password lama.'
                  example: password123
                new_password:
                  type: string
                  description: 'Password baru minimal 8 karakter.'
                  example: password456
                new_password_confirmation:
                  type: string
                  description: 'Konfirmasi password baru.'
                  example: password456
              required:
                - old_password
                - new_password
                - new_password_confirmation
...